ODE-1006 patch
I have been trying to use the Management Console of ODE for a while, but always failed to do so because none of the buttons were working. I started looking at the code and it seems really broken (innerHTML with <script> tags, etc.) so I took a stab at trying to make it run under Firefox at least.

The console doesn't work in Chrome still, mostly because of services not returning the correct 'Content-Type' of 'text/xml' and thus having no access to '.responseXML' in JavaScript.